+++ ddprof-lib/src/main/cpp/threadFilter.h
@@ -45,17 +45,25 @@ public:
     bool enabled() const;
     // Hot path methods - slot_id MUST be from registerThread(), undefined behavior otherwise
     bool accept(SlotID slot_id) const;
-    void add(long tid, SlotID slot_id);
+    void add(long java_tid, int os_tid, SlotID slot_id);
     void remove(SlotID slot_id);
-    void collect(std::vector<int>& tids) const;
+
+    void collect_java_tids(std::vector<long>& java_tids) const;
+    void collect_os_tids(std::vector<int>& os_tids) const;
 
     SlotID registerThread();
     void unregisterThread(SlotID slot_id);
 
 private:
+    // Use os tid for filtering vs java tid
+    static bool filter_by_os_tid();
+
     // Optimized slot structure with padding to avoid false sharing
     struct alignas(DEFAULT_CACHE_LINE_SIZE) Slot {
-        std::atomic<long> value{-1};
+        union {
+            std::atomic<long> java_tid;
+            std::atomic<int>  os_tid;
+        } value;
         char padding[DEFAULT_CACHE_LINE_SIZE - sizeof(value)];  // Pad to cache line size
     };
     static_assert(sizeof(Slot) == DEFAULT_CACHE_LINE_SIZE, "Slot must be exactly one cache line");
