+++ ddprof-lib/src/main/cpp/wallclock/wallClock.cpp
@@ -1,0 +1,62 @@
+/*
+ * Copyright The async-profiler authors
+ * Copyright 2025, Datadog, Inc.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#include "wallClock.h"
+#include "stackFrame.h"
+#include "context.h"
+#include "debugSupport.h"
+#include "libraries.h"
+#include "log.h"
+#include "profiler.h"
+#include "stackFrame.h"
+#include "thread.h"
+#include "threadState.inline.h"
+#include "vmStructs.h"
+#include "guards.h"
+#include <math.h>
+#include <random>
+#include <algorithm> // For std::sort and std::binary_search
+
+std::atomic<bool> BaseWallClock::_enabled{false};
+
+Error BaseWallClock::start(Arguments &args) {
+  int interval = args._event != NULL ? args._interval : args._wall;
+  if (interval < 0) {
+    return Error("interval must be positive");
+  }
+  _interval = interval ? interval : DEFAULT_WALL_INTERVAL;
+
+    _reservoir_size =
+            args._wall_threads_per_tick ?
+            args._wall_threads_per_tick
+                                                : DEFAULT_WALL_THREADS_PER_TICK;
+
+  initialize(args);
+
+  _running = true;
+
+  if (pthread_create(&_thread, NULL, threadEntry, this) != 0) {
+    return Error("Unable to create timer thread");
+  }
+
+  return Error::OK;
+}
+
+void BaseWallClock::stop() {
+  _running.store(false);
+  // the thread join ensures we wait for the thread to finish before returning
+  // (and possibly removing the object)
+  pthread_kill(_thread, WAKEUP_SIGNAL);
+  int res = pthread_join(_thread, NULL);
+  if (res != 0) {
+    Log::warn("Unable to join WallClock thread on stop %d", res);
+  }
+}
+
+bool BaseWallClock::isEnabled() const {
+  return _enabled.load(std::memory_order_acquire);
+}
+
