+++ ddprof-lib/src/main/cpp/wallclock/wallClockASGCT.cpp
@@ -1,0 +1,151 @@
+/*
+ * Copyright The async-profiler authors
+ * Copyright 2026, Datadog, Inc.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#include "debugSupport.h"
+#include "guards.h"
+#include "stackFrame.h"
+#include "threadState.inline.h"
+#include "wallclock/wallClock.inline.h"
+#include "wallclock/wallClockASGCT.h"
+
+bool WallClockASGCT::inSyscall(void *ucontext) {
+  StackFrame frame(ucontext);
+  uintptr_t pc = frame.pc();
+
+  // Consider a thread sleeping, if it has been interrupted in the middle of
+  // syscall execution, either when PC points to the syscall instruction, or if
+  // syscall has just returned with EINTR
+  if (StackFrame::isSyscall((instruction_t *)pc)) {
+    return true;
+  }
+
+  // Make sure the previous instruction address is readable
+  uintptr_t prev_pc = pc - SYSCALL_SIZE;
+  if ((pc & 0xfff) >= SYSCALL_SIZE ||
+      Libraries::instance()->findLibraryByAddress((instruction_t *)prev_pc) !=
+          NULL) {
+    if (StackFrame::isSyscall((instruction_t *)prev_pc) &&
+        frame.checkInterruptedSyscall()) {
+      return true;
+    }
+  }
+
+  return false;
+}
+
+void WallClockASGCT::sharedSignalHandler(int signo, siginfo_t *siginfo,
+                                    void *ucontext) {
+  WallClockASGCT *engine = reinterpret_cast<WallClockASGCT *>(Profiler::instance()->wallEngine());
+  if (signo == SIGVTALRM) {
+    engine->signalHandler(signo, siginfo, ucontext, engine->_interval);
+  }
+}
+
+void WallClockASGCT::signalHandler(int signo, siginfo_t *siginfo, void *ucontext,
+                              u64 last_sample) {
+  // Atomically try to enter critical section - prevents all reentrancy races
+  CriticalSection cs;
+  if (!cs.entered()) {
+    return;  // Another critical section is active, defer profiling
+  }
+  ProfiledThread *current = ProfiledThread::currentSignalSafe();
+  int tid = current != NULL ? current->tid() : OS::threadId();
+  Shims::instance().setSighandlerTid(tid);
+  u64 call_trace_id = 0;
+  if (current != NULL && _collapsing) {
+    StackFrame frame(ucontext);
+    Context &context = Contexts::get();
+    call_trace_id = current->lookupWallclockCallTraceId(
+        (u64)frame.pc(), (u64)frame.sp(),
+        Profiler::instance()->recordingEpoch(),
+        context.spanId, context.rootSpanId);
+    if (call_trace_id != 0) {
+      Counters::increment(SKIPPED_WALLCLOCK_UNWINDS);
+    }
+  }
+
+  ExecutionEvent event;
+  VMThread *vm_thread = VMThread::current();
+  if (vm_thread != NULL && !vm_thread->isThreadAccessible()) {
+      vm_thread = NULL;
+  }
+  int raw_thread_state = vm_thread ? vm_thread->state() : 0;
+  bool is_java_thread = raw_thread_state >= 4 && raw_thread_state < 12;
+  bool is_initialized = is_java_thread;
+  OSThreadState state = OSThreadState::UNKNOWN;
+  ExecutionMode mode = ExecutionMode::UNKNOWN;
+  if (vm_thread && is_initialized) {
+    OSThreadState os_state = vm_thread->osThreadState();
+    if (os_state != OSThreadState::UNKNOWN) {
+      state = os_state;
+    }
+    mode = getThreadExecutionMode();
+  }
+  if (state == OSThreadState::UNKNOWN) {
+    if (inSyscall(ucontext)) {
+      state = OSThreadState::SYSCALL;
+      mode = ExecutionMode::SYSCALL;
+    } else {
+      state = OSThreadState::RUNNABLE;
+    }
+  }
+  event._thread_state = state;
+  event._execution_mode = mode;
+  event._weight = 1;
+  Profiler::instance()->recordSample(ucontext, last_sample, tid, BCI_WALL,
+                                     call_trace_id, &event);
+  Shims::instance().setSighandlerTid(-1);
+}
+
+
+void WallClockASGCT::timerLoop() {
+    // todo: re-allocating the vector every time is not efficient
+    auto collectThreads = [&](std::vector<int>& tids) {
+      // Get thread IDs from the filter if it's enabled
+      // Otherwise list all threads in the system
+      if (Profiler::instance()->threadFilter()->enabled()) {
+        Profiler::instance()->threadFilter()->collect_os_tids(tids);
+      } else {
+        ThreadList *thread_list = OS::listThreads();
+        while (thread_list->hasNext()) {
+          int tid = thread_list->next();
+          // Don't include the current thread
+          if (tid != OS::threadId()) {
+            tids.push_back(tid);
+          }
+          tid = thread_list->next();
+        }
+        delete thread_list;
+      }
+    };
+
+    auto sampleThreads = [&](int tid, int& num_failures, int& threads_already_exited, int& permission_denied) {
+      if (!OS::sendSignalToThread(tid, SIGVTALRM)) {
+        num_failures++;
+        if (errno != 0) {
+          if (errno == ESRCH) {
+              threads_already_exited++;
+          } else if (errno == EPERM) {
+              permission_denied++;
+          } else {
+              Log::debug("unexpected error %s", strerror(errno));
+          }
+        }
+        return false;
+      }
+      return true;
+    };
+
+    auto doNothing = []() {
+    };
+
+    timerLoopCommon<int>(collectThreads, sampleThreads, doNothing, _reservoir_size, _interval);
+}
+
+void WallClockASGCT::initialize(Arguments& args) {
+  _collapsing = args._wall_collapsing;
+  OS::installSignalHandler(SIGVTALRM, sharedSignalHandler);
+}
